+++ bin/main.rb
@@ -3,8 +3,82 @@ require 'nokogiri'
 require 'open-uri'
 require_relative '../lib/result'
 require_relative '../lib/validation'
-require_relative '../lib/search'
-require_relative '../lib/input'
+require_relative '../lib/link'
+
+class Input
+  def initialize(link, validation)
+    @link = link
+    @validation = validation
+    @term = ''
+    @inclusion = ''
+    @do_again = ''
+  end
+
+  def input_entry
+    loop do
+      puts
+      card_type_entry
+      puts
+      inclusion_entry
+      puts
+      do_again
+      puts
+      @link.build_link(@term, @inclusion, @validation.current_term)
+      puts @link.chosen_terms
+      break if @repeat == 'N'
+    end
+  end
+
+  private
+
+  def card_type_entry
+    loop do
+      print "Chose a Card #{@validation.current_term}: "
+      @term = gets.chomp.downcase.capitalize
+      break if @validation.type_array.include?(@term) || @validation.current_term == 'Subtype'
+
+      puts "Not a valid Card #{@validation.current_term}. "
+    end
+  end
+
+  def inclusion_entry
+    loop do
+      print 'Include or Exclude from the search? (I/E): '
+      @inclusion = gets.chomp.upcase
+      break if @validation.inclusion_validation.include? @inclusion
+
+      puts 'Invalid Input. Chose either I or N'
+    end
+  end
+
+  def do_again
+    loop do
+      print 'Want to add another? (Y/N): '
+      @repeat = gets.chomp.upcase
+      break if @validation.repeat_validation.include? @repeat
+
+      puts 'Invalid Input. Chose either Y or N'
+    end
+  end
+end
+
+def show_result(result, validation)
+  puts "Your search returned #{result.length} results."
+  puts
+rescue Errno::ETIMEDOUT
+  puts 'Site is down. Try again in a couple minutes.'
+else
+  list_result = 'N'
+  loop do
+    print 'Want to view a list with the results? (Y/N): '
+    list_result = gets.chomp.upcase
+    break if validation.include? list_result
+
+    puts 'Invalid Input. Chose either Y or N'
+  end
+  puts
+  puts result if list_result == 'Y'
+end
 
 puts
 puts '------------ Welcome to the Magic: The Gathering Web Scraper ------------'
@@ -26,10 +100,10 @@ puts ' This is not mandatory and accepts any value.'
 puts
 puts '-------------------------------------------------------------------------'
 
-search = Search.new
+link = Link.new
 validation = Validation.new
 result = Result.new
-input = Input.new(search, validation)
+input = Input.new(link, validation)
 puts
 puts "The possible Types or Supertypes are: #{validation.type_array}"
 puts
@@ -38,5 +112,5 @@ input.input_entry
 validation.change_term
 input.input_entry
 puts
-search.web_scrape
-result.show_result(search.name_array, validation.repeat_validation)
+result.web_scrape(link.base_link, link.options_link, link.type_link, link.subtype_link)
+show_result(result.name_array, validation.repeat_validation)
