+++ lib/result.rb
@@ -3,21 +3,30 @@ require 'nokogiri'
 require 'open-uri'
 
 class Result
-  def show_result(result, validation)
-    puts "Your search returned #{result.length} results."
-    puts
-  rescue Errno::ETIMEDOUT
-    puts 'Site is down. Try again in a couple minutes.'
-  else
-    list_result = 'N'
-    loop do
-      print 'Want to view a list with the results? (Y/N): '
-      list_result = gets.chomp.upcase
-      break if validation.include? list_result
+  attr_reader :name_array
 
-      puts 'Invalid Input. Chose either Y or N'
+  def initialize
+    @page = 0
+    @name_array = []
+    @page_array = []
+  end
+
+  def web_scrape(base_link, options_link, type_link, subtype_link)
+    loop do
+      full_link = "#{base_link}page=#{@page}#{options_link}#{type_link}#{subtype_link}"
+      retrieved_page = Nokogiri::HTML(URI.parse(full_link).open)
+      retrieved_page.xpath('//td//a').each do |content|
+        @name_array << content.content unless content.content == ''
+      end
+
+      retrieved_page.css('#ctl00_ctl00_ctl00_MainContent_SubContent_topPagingControlsContainer a').each do |link|
+        @page_array << link.content.to_i - 1 unless link.content.to_i.zero?
+      end
+
+      break if @page == @page_array[-1] or @page_array.empty?
+
+      @page += 1
     end
-    puts
-    puts result if list_result == 'Y'
+    @name_array
   end
 end
